refactor(traces): consolidate Selection/SelectionMarker into common
The scatter_geo, scatter_map, and scatter_mapbox modules each defined a
byte-identical Selection/SelectionMarker pair. Move the single definition
to crate::common and re-export it from each module (pub use) so the public
paths plotly::traces::scatter_geo::Selection etc. keep resolving — no
serialized-output change, all existing trace tests pass unchanged.

This is the shared building block for backfilling selected/unselected
styling onto the older cartesian traces.
