+++ plotly/src/traces/scatter_mapbox.rs
@@ -17,42 +17,10 @@ pub enum Fill {
     ToSelf,
 }
 
-#[serde_with::skip_serializing_none]
-#[derive(Serialize, Clone, Debug, Default)]
-pub struct SelectionMarker {
-    color: Option<Box<dyn Color>>,
-    opacity: Option<f64>,
-    size: Option<Dim<usize>>,
-}
-
-#[derive(Serialize, Clone, Debug, Default)]
-pub struct Selection {
-    marker: SelectionMarker,
-}
-
-impl Selection {
-    pub fn new() -> Self {
-        Default::default()
-    }
-
-    /// Sets the marker color of un/selected points.
-    pub fn color<C: Color>(mut self, color: C) -> Self {
-        self.marker.color = Some(Box::new(color));
-        self
-    }
-
-    /// Sets the marker opacity of un/selected points.
-    pub fn opacity(mut self, opacity: f64) -> Self {
-        self.marker.opacity = Some(opacity);
-        self
-    }
-
-    /// Sets the marker size of un/selected points.
-    pub fn size(mut self, size: usize) -> Self {
-        self.marker.size = Some(Dim::Scalar(size));
-        self
-    }
-}
+// `Selection` and `SelectionMarker` are defined in `crate::common` and
+// re-exported here for backward compatibility (they used to live in this
+// module).
+pub use crate::common::{Selection, SelectionMarker};
 
 #[serde_with::skip_serializing_none]
 #[derive(Serialize, Clone, Debug, FieldSetter)]
